# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.CategoryResponse:
    properties:
      date_created:
        description: |-
          Дата создания
          example: 02.01.2006
        type: string
      id:
        description: |-
          ID категории
          example: 1
        type: number
      img_url:
        description: |-
          URL изображения категории
          example: https://example.com/category.jpg
        type: string
      name:
        description: |-
          Название категории
          example: Утренние практики
        type: string
      types:
        description: Список типов категории
        items:
          $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.TypeResponse'
        type: array
    type: object
  github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse:
    properties:
      details:
        description: |-
          Детали ошибки (опционально)
          example: Login field is required
        type: string
      error:
        description: |-
          Текст ошибки
          example: Invalid request
        type: string
    type: object
...
  github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.TypeResponse:
    properties:
      created_at:
        description: |-
          Дата создания
          example: 02.01.2006
        type: string
      id:
        description: |-
          ID типа
          example: 1
        type: number
      name:
        description: |-
          Название типа
          example: Йога
        type: string
    type: object
...
  github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.VideoRequest:
    properties:
      category_ids:
        description: |-
          Список ID категорий
          example: [1, 2, 3]
        items:
          type: integer
        type: array
      description:
        description: |-
          Описание видео
          example: 30-минутный комплекс утренних упражнений
        type: string
      img_url:
        description: |-
          URL превью изображения
          example: https://example.com/preview.jpg
        type: string
      name:
        description: |-
          Название видео
          required: true
          example: Утренняя йога
        type: string
      url:
        description: |-
          URL видеофайла
          required: true
          example: https://example.com/video.mp4
        type: string
    type: object
  github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.VideoResponse:
    properties:
      categories:
        description: Список категорий видео
        items:
          $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.CategoryResponse'
        type: array
      created_at:
        description: |-
          Дата создания
          example: 02.01.2006
        type: string
      description:
        description: |-
          Описание видео
          example: 30-минутный комплекс утренних упражнений
        type: string
      id:
        description: |-
          ID видео
          example: 1
        type: number
      img_url:
        description: |-
          URL превью изображения
          example: https://example.com/preview.jpg
        type: string
      name:
        description: |-
          Название видео
          example: Утренняя йога
        type: string
      url:
        description: |-
          URL видеофайла
          example: https://example.com/video.mp4
        type: string
    type: object
...
  github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.CategoryResponse:
    description: Информация о категории контента
    properties:
      id:
        description: ID из БД
        type: number
      img_url:
        description: Превью картинка для категории
        type: string
      name:
        description: Название категории
        type: string
    type: object
  github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.ErrorResponse:
    description: Стандартная структура для возврата ошибок
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.VideoResponse:
    description: Информация о видео, включая URL, описание и категорию
    properties:
      category:
        description: Название категории
        type: string
      description:
        description: Описание видео
        type: string
      id:
        description: ID из БД
        type: number
      img_url:
        description: Превью картинка для видео
        type: string
      name:
        description: Название видео
        type: string
      url:
        description: URL адрес до файла
        type: string
    type: object
host: https://api.7375.org
info:
  contact:
    email: info@7375.org
    name: Sergei
  description: API для управления видео-контентом BodyBalance.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: BodyBalance API
  version: "1.0"
paths:
...
  /admin/video:
    get:
      description: Возвращает список всех видео в системе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.VideoResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Получить список всех видео
      tags:
      - Admin Videos
    post:
      consumes:
      - application/json
      description: Создает новую запись видео в системе
      parameters:
      - description: Данные видео
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.VideoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Видео успешно добавлено
          schema:
            properties:
              id:
                type: integer
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Добавить новое видео
      tags:
      - Admin Videos
  /admin/video/{id}:
    delete:
      description: Помечает видео как удаленное в системе
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Видео успешно удалено
          schema:
            properties:
              id:
                type: integer
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Удалить видео
      tags:
      - Admin Videos
    get:
      description: Возвращает информацию о конкретном видео
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.VideoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Получить видео по ID
      tags:
      - Admin Videos
    put:
      consumes:
      - application/json
      description: Обновляет информацию о существующем видео
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные видео
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.VideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Видео успешно обновлено
          schema:
            properties:
              id:
                type: integer
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Обновить данные видео
      tags:
      - Admin Videos
  /category:
    get:
      description: Returns all categories for specified type, ordered by name
      parameters:
      - description: Type ID
        in: query
        name: type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.CategoryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.ErrorResponse'
      summary: Get categories by type
      tags:
      - API v1
...
  /video:
    get:
      description: Returns video details by its ID
      parameters:
      - description: Video ID
        in: query
        name: video_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.VideoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.ErrorResponse'
      summary: Get video by ID
      tags:
      - API v1
...
  /video_categories:
    get:
      description: Returns videos filtered by type and category, ordered by name
      parameters:
      - description: Type ID
        in: query
        name: type
        required: true
        type: integer
      - description: Category ID
        in: query
        name: category
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.VideoResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_api_v1_response.ErrorResponse'
      summary: Get videos by category and type
      tags:
      - API v1
swagger: "2.0"