# internal/http-server/admin/video.go
package admin

import (
	"database/sql"
	"encoding/json"
	"errors"
	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/langowen/bodybalance-backend/internal/http-server/admin/admResponse"
	"github.com/langowen/bodybalance-backend/internal/lib/logger/sl"
	"net/http"
	"strconv"
)

// @Summary Добавить новое видео
// @Description Создает новую запись видео в системе
// @Tags Admin Videos
// @Accept json
// @Produce json
// @Param input body admResponse.VideoRequest true "Данные видео"
// @Success 201 {object} object{id=int64,message=string} "Видео успешно добавлено"
// @Failure 400 {object} admResponse.ErrorResponse
// @Failure 500 {object} admResponse.ErrorResponse
// @security AdminAuth
// @Router /admin/video [post]
func (h *Handler) addVideo(w http.ResponseWriter, r *http.Request) {
	const op = "admin.addVideo"

	logger := h.logger.With(
		"op", op,
		"request_id", middleware.GetReqID(r.Context()),
	)

	var req admResponse.VideoRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		logger.Error("failed to decode request body", sl.Err(err))
		admResponse.RespondWithError(w, http.StatusBadRequest, "Invalid request format")
		return
	}

	if req.URL == "" || req.Name == "" {
		logger.Error("empty required fields")
		admResponse.RespondWithError(w, http.StatusBadRequest, "URL and Name are required")
		return
	}

	ctx := r.Context()
	videoID, err := h.storage.AddVideo(ctx, req)
	if err != nil {
		logger.Error("failed to add video", sl.Err(err))
		admResponse.RespondWithError(w, http.StatusInternalServerError, "Failed to add video")
		return
	}

	// Добавляем связи с категориями
	if len(req.CategoryIDs) > 0 {
		err = h.storage.AddVideoCategories(ctx, videoID, req.CategoryIDs)
		if err != nil {
			logger.Error("failed to add video categories", sl.Err(err))
			admResponse.RespondWithError(w, http.StatusInternalServerError, "Failed to add video categories")
			return
		}
	}

	admResponse.RespondWithJSON(w, http.StatusCreated, map[string]interface{}{
		"id":      videoID,
		"message": "Video added successfully",
	})
}

// @Summary Получить видео по ID
// @Description Возвращает информацию о конкретном видео
// @Tags Admin Videos
// @Produce json
// @Param id path int true "ID видео"
// @Success 200 {object} admResponse.VideoResponse
// @Failure 400 {object} admResponse.ErrorResponse
// @Failure 404 {object} admResponse.ErrorResponse
// @Failure 500 {object} admResponse.ErrorResponse
// @security AdminAuth
// @Router /admin/video/{id} [get]
func (h *Handler) getVideo(w http.ResponseWriter, r *http.Request) {
	const op = "admin.getVideo"

	logger := h.logger.With(
		"op", op,
		"request_id", middleware.GetReqID(r.Context()),
	)

	idStr := chi.URLParam(r, "id")
	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		logger.Error("invalid video ID", sl.Err(err))
		admResponse.RespondWithError(w, http.StatusBadRequest, "Invalid video ID")
		return
	}

	ctx := r.Context()
	video, err := h.storage.GetVideo(ctx, id)
	if err != nil {
		if errors.Is(err, sql.ErrNoRows) {
			logger.Warn("video not found", "video_id", id)
			admResponse.RespondWithError(w, http.StatusNotFound, "Video not found")
			return
		}
		logger.Error("failed to get video", sl.Err(err), "video_id", id)
		admResponse.RespondWithError(w, http.StatusInternalServerError, "Failed to get video")
		return
	}

	// Получаем категории видео
	categories, err := h.storage.GetVideoCategories(ctx, id)
	if err != nil {
		logger.Error("failed to get video categories", sl.Err(err), "video_id", id)
		admResponse.RespondWithError(w, http.StatusInternalServerError, "Failed to get video categories")
		return
	}

	video.Categories = categories

	admResponse.RespondWithJSON(w, http.StatusOK, video)
}

// @Summary Получить список всех видео
// @Description Возвращает список всех видео в системе
// @Tags Admin Videos
// @Produce json
// @Success 200 {array} admResponse.VideoResponse
// @Failure 500 {object} admResponse.ErrorResponse
// @security AdminAuth
// @Router /admin/video [get]
func (h *Handler) getVideos(w http.ResponseWriter, r *http.Request) {
	const op = "admin.getVideos"

	logger := h.logger.With(
		"op", op,
		"request_id", middleware.GetReqID(r.Context()),
	)

	ctx := r.Context()
	videos, err := h.storage.GetVideos(ctx)
	if err != nil {
		logger.Error("failed to get videos", sl.Err(err))
		admResponse.RespondWithError(w, http.StatusInternalServerError, "Failed to get videos")
		return
	}

	// Для каждого видео получаем категории
	for i := range videos {
		categories, err := h.storage.GetVideoCategories(ctx, int64(videos[i].ID))
		if err != nil {
			logger.Error("failed to get video categories", sl.Err(err), "video_id", videos[i].ID)
			admResponse.RespondWithError(w, http.StatusInternalServerError, "Failed to get video categories")
			return
		}
		videos[i].Categories = categories
	}

	admResponse.RespondWithJSON(w, http.StatusOK, videos)
}

// @Summary Обновить данные видео
// @Description Обновляет информацию о существующем видео
// @Tags Admin Videos
// @Accept json
// @Produce json
// @Param id path int true "ID видео"
// @Param input body admResponse.VideoRequest true "Новые данные видео"
// @Success 200 {object} object{id=int64,message=string} "Видео успешно обновлено"
// @Failure 400 {object} admResponse.ErrorResponse
// @Failure 404 {object} admResponse.ErrorResponse
// @Failure 500 {object} admResponse.ErrorResponse
// @security AdminAuth
// @Router /admin/video/{id} [put]
func (h *Handler) updateVideo(w http.ResponseWriter, r *http.Request) {
	const op = "admin.updateVideo"

	logger := h.logger.With(
		"op", op,
		"request_id", middleware.GetReqID(r.Context()),
	)

	idStr := chi.URLParam(r, "id")
	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		logger.Error("invalid video ID", sl.Err(err))
		admResponse.RespondWithError(w, http.StatusBadRequest, "Invalid video ID")
		return
	}

	var req admResponse.VideoRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		logger.Error("failed to decode request body", sl.Err(err))
		admResponse.RespondWithError(w, http.StatusBadRequest, "Invalid request format")
		return
	}

	ctx := r.Context()
	err = h.storage.UpdateVideo(ctx, id, req)
	if err != nil {
		if errors.Is(err, sql.ErrNoRows) {
			logger.Warn("video not found", "video_id", id)
			admResponse.RespondWithError(w, http.StatusNotFound, "Video not found")
			return
		}
		logger.Error("failed to update video", sl.Err(err), "video_id", id)
		admResponse.RespondWithError(w, http.StatusInternalServerError, "Failed to update video")
		return
	}

	// Обновляем категории видео
	if len(req.CategoryIDs) > 0 {
		// Сначала удаляем все существующие связи
		err = h.storage.DeleteVideoCategories(ctx, id)
		if err != nil {
			logger.Error("failed to delete video categories", sl.Err(err), "video_id", id)
			admResponse.RespondWithError(w, http.StatusInternalServerError, "Failed to update video categories")
			return
		}

		// Затем добавляем новые
		err = h.storage.AddVideoCategories(ctx, id, req.CategoryIDs)
		if err != nil {
			logger.Error("failed to add video categories", sl.Err(err), "video_id", id)
			admResponse.RespondWithError(w, http.StatusInternalServerError, "Failed to update video categories")
			return
		}
	}

	admResponse.RespondWithJSON(w, http.StatusOK, map[string]interface{}{
		"id":      id,
		"message": "Video updated successfully",
	})
}

// @Summary Удалить видео
// @Description Помечает видео как удаленное в системе
// @Tags Admin Videos
// @Produce json
// @Param id path int true "ID видео"
// @Success 200 {object} object{id=int64,message=string} "Видео успешно удалено"
// @Failure 400 {object} admResponse.ErrorResponse
// @Failure 404 {object} admResponse.ErrorResponse
// @Failure 500 {object} admResponse.ErrorResponse
// @security AdminAuth
// @Router /admin/video/{id} [delete]
func (h *Handler) deleteVideo(w http.ResponseWriter, r *http.Request) {
	const op = "admin.deleteVideo"

	logger := h.logger.With(
		"op", op,
		"request_id", middleware.GetReqID(r.Context()),
	)

	idStr := chi.URLParam(r, "id")
	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		logger.Error("invalid video ID", sl.Err(err))
		admResponse.RespondWithError(w, http.StatusBadRequest, "Invalid video ID")
		return
	}

	ctx := r.Context()
	err = h.storage.DeleteVideo(ctx, id)
	if err != nil {
		if errors.Is(err, sql.ErrNoRows) {
			logger.Warn("video not found", "video_id", id)
			admResponse.RespondWithError(w, http.StatusNotFound, "Video not found")
			return
		}
		logger.Error("failed to delete video", sl.Err(err), "video_id", id)
		admResponse.RespondWithError(w, http.StatusInternalServerError, "Failed to delete video")
		return
	}

	admResponse.RespondWithJSON(w, http.StatusOK, map[string]interface{}{
		"id":      id,
		"message": "Video deleted successfully",
	})
}